+++ prove.sh
@@ -12,7 +12,7 @@ else  # Linux
   for PY_VER in 3.9 3.10 3.11 3.12; do
     PYTHON="/opt/python$PY_VER/bin/python3"
     echo "===== Running" "$PYTHON" "====="
-    "$PYTHON" dev/gencovrc.py -q -f$PY_VER 9 13
+    "$PYTHON" -c 'from igbpyutils.dev import generate_coveragerc_cli as main; main()' -q -f$PY_VER 9 13
     COVERAGE="/opt/python$PY_VER/bin/coverage"
     PYTHONWARNINGS=error PYTHONWARNDEFAULTENCODING=1 "$COVERAGE" run --rcfile=".coveragerc$PY_VER" --branch -m unittest "$@"
     "$COVERAGE" report --rcfile=".coveragerc$PY_VER" --omit='*/igbdatatools/*' --skip-covered --show-missing --fail-under=100 | grep -v 'skipped due to complete coverage'