+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 v0.4.0 - Fri Dec  8 2023
     - Changed :func:`~igbpyutils.file.autoglob` to detect current shell
       via environment variables instead of ``sys.platform``
+    - Added :func:`igbpyutils.dev.generate_coveragerc` and CLI
 
 v0.3.2 - Thu Nov  2 2023
     - Added ``--add*`` options to ``simple-perms`` tool